pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "stagger"
version = "0.1.0"
description = "Encoder-only protein tagger (SDPA + RoPE + SwiGLU)"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [{ name = "Bryan Briney" }]
dependencies = [
  "torch>=2.1",
  "hydra-core>=1.3",
  "omegaconf>=2.3",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["stagger*"]
exclude = []


